experimento-2/gateway/requirements.txt
flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
pybreaker==1.2.0
PyJWT==2.8.0